python/floorplan-ai/requirements.txt
# Core ML
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0

# OCR
easyocr>=1.7.0

# Data Processing
numpy>=1.24.0
scipy>=1.11.0
shapely>=2.0.0

# API Server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# SVG Generation
svgwrite>=1.4.3

# Config & Logging
pyyaml>=6.0
loguru>=0.7.0

# PDF Processing
pdf2image>=1.16.0
PyMuPDF>=1.23.0

# Dev & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
